--- metagpt/environment/mincraft_env/mineflayer/.gitignore
@@ -1,1 +1,0 @@
-!/lib
