--- metagpt/environment/mincraft_env/mineflayer/.prettierignore
@@ -1,3 +1,0 @@
-# Ignore artifacts:
-build
-coverage
