--- metagpt/environment/mincraft_env/mineflayer/.prettierrc.json
@@ -1,3 +1,0 @@
-{
-    "tabWidth": 4
-}
