--- metagpt/environment/mincraft_env/mineflayer/mineflayer-collectblock/_config.yml
@@ -1,1 +1,0 @@
-theme: jekyll-theme-cayman
