--- metagpt/environment/stanford_town_env/__init__.py
@@ -1,3 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# @Desc   :
