--- metagpt/environment/stanford_town_env/stanford_town_env.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# @Desc   : MG StanfordTown Env
-
-from metagpt.environment.stanford_town_env.stanford_town_ext_env import (
-    StanfordTownExtEnv,
-)
-
-
-class StanfordTownEnv(StanfordTownExtEnv):
-    pass
